# Optimize position search in error path