@purplesyringa 2024-07-26
For easier comparison, here are the results of the benchmark for the success path on the same machine:
```
                        DOM         STRUCT
data/canada.json        283 MB/s    416 MB/s
data/citm_catalog.json  429 MB/s    864 MB/s
data/twitter.json       275 MB/s    541 MB/s
```